server/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "schedulo"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "Django==5.1.3",
    "django-cors-headers==4.6.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "gunicorn==23.0.0",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.1.0",
    "PyJWT==2.10.1",
    "sqlparse==0.5.2",
    "whitenoise==6.9.0",
]

[project.optional-dependencies]
dev = [
    "black==25.1.0",
    "flake8==7.2.0",
]

[tool.setuptools.packages.find]
where = ["server"]

[tool.black]
line-length = 88
exclude = '''
/(
      \.eggs
      | build
      | venv
      | \.venv
      | \.git
      | node_modules
      | __pycache__
      | \.vscode
      | \.vite
      | dist
)/
'''
